Create blockchain::network.rs and blockchain::storage.rs to handle all related structs and methods (#352)
* Create network.rs and storage.rs to handle related methods

* Add parse genesis block function

Remove unused tests data

* Update src/blockchain/src/network.rs


* Change MyBehaviour ro Behaviour
