+++ src/blockchain/src/lib.rs
@@ -17,3 +17,5 @@
 pub mod block;
 pub mod blockchain;
 pub mod header;
+pub mod network;
+pub mod storage;
